Chapter 9: Dielectrics (final pass)
Unfortunately the previous fix for the cosine computation introduces NaNs. That's because the fix involves the square root of a discriminant which can be negative. When it's negative it means that there's total internal reflection.
At the same time rearrange the dielectric scattering implementation to avoid the premature computation of the cosine value.